maint: tr: tweak/make-consistent a pair of loop indices
* src/tr.c (unquote): Use indices of type "unsigned int", rather
than a mix or "size_t" and "int", presuming that no command-line-
specified string will have length longer than UINT_MAX.

+++ src/tr.c
@@ -436,8 +436,8 @@ unquote (char const *s, struct E_string *es)
   es->s = xmalloc (len);
   es->escaped = xcalloc (len, sizeof es->escaped[0]);
 
-  size_t j = 0;
-  for (int i = 0; s[i]; i++)
+  unsigned int j = 0;
+  for (unsigned int i = 0; s[i]; i++)
     {
       unsigned char c;
       int oct_digit;
